insight_face/app/requirements.txt
opencv-python
insightface
onnxruntime
uvicorn[standard]
pillow
numpy
python-multipart
fastapi==0.98.0
